+++ .gitignore
@@ -1,7 +1,1 @@
-target
 */target
-*.bak
-*/*.bak
-CMakeProject1
-*/CMakeProject1
-mysql.log
