+++ .idea/encodings.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/eureka_server/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/order_sertvice/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/order_sertvice/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/product_service/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+  <component name="Encoding" addBOMForNewFiles="with NO BOM">
+    <file url="file://$PROJECT_DIR$" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/eureka_server" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/order_service" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/product_service" charset="UTF-8" />
   </component>
 </project>
