--- order_sertvice/src/main/resources/application.yml
@@ -1,17 +1,0 @@
-server:
-  port: 9002
-  servlet:
-    context-path: /order
-
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/shop?characterEncoding=utf-8&serverTimezone=UTC&useSSL=false
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    username: root
-    password: 1234
-  application:
-    name: order_service
-  jpa:
-    database: MySQL
-    show-sql: true
-    open-in-view: true
