+++ order_service/src/main/java/com/huike/order/OrderApplication.java
@@ -1,0 +1,23 @@
+package com.huike.order;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
+
+@SpringBootApplication
+//@EntityScan("com.huike.order.entity")
+
+public class OrderApplication {
+    @LoadBalanced
+    @Bean
+    public RestTemplate restTemplate(){
+        return  new RestTemplate();
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(OrderApplication.class,args);
+    }
+}
