+++ order_service/src/main/java/com/huike/order/controller/OrderController.java
@@ -1,0 +1,30 @@
+package com.huike.order.controller;
+
+import com.huike.order.entity.Product;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.discovery.DiscoveryClient;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/order")
+public class OrderController {
+
+    //    注入restTemplate对象，实现微服务间调用
+    @Autowired
+    private RestTemplate restTemplate;
+
+    @Autowired
+    private DiscoveryClient discoveryClient;
+
+    @RequestMapping(value = "/buy/{id}", method = RequestMethod.GET)
+    public Product findById(@PathVariable Long id) {
+        return restTemplate.getForObject("http://service-product/product/" + id, Product.class);
+    }
+}
