+++ order_service/src/main/java/com/huike/order/entity/Product.java
@@ -1,17 +1,11 @@
 package com.huike.order.entity;
 
 import lombok.Data;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import java.math.BigDecimal;
 
 @Data
-@Entity
-@Table(name = "tb_product")
 public class Product {
-    @Id
+
     private Long id;
     private String productName;
     private Integer status;
