+++ order_service/src/main/resources/application.yml
@@ -1,0 +1,11 @@
+server:
+  port: 9002 #端口
+spring:
+  application:
+    name: service-order #服务名称
+eureka:
+  client:
+    serviceUrl:
+      defaultZone: http://localhost:9000/eureka/
+  instance:
+    prefer-ip-address: true
