+++ product_service/src/main/java/com/huike/product/controller/ProductController.java
@@ -3,28 +3,21 @@ package com.huike.product.controller;
 import com.huike.product.entity.Product;
 import com.huike.product.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("/product")
-//允许跨域访问
-@CrossOrigin(origins = "*")
 @RestController
+@RequestMapping("/product")
 public class ProductController {
 
     @Autowired
     private ProductService productService;
 
-    @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-
-    public Product findById(@PathVariable Long id) {
-
-        try {
-            return productService.findById(id);
-        } catch (Exception e) {
-            System.out.println("有异常");
-        }
-        return null;
+    @RequestMapping(value="/{id}",method= RequestMethod.GET)
+    public Product findById(@PathVariable Long id){
+        Product product = productService.findById(id);
+        return  product;
     }
-
-
 }
