+++ product_service/src/main/java/com/huike/product/dao/ProductDao.java
@@ -3,8 +3,6 @@ package com.huike.product.dao;
 import com.huike.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.apache.ibatis.annotations.*;
-public interface ProductDao extends JpaRepository <Product,Long>, JpaSpecificationExecutor<Product> {
 
-
+public interface ProductDao extends JpaRepository<Product,Long>, JpaSpecificationExecutor<Product> {
 }
