+++ product_service/src/main/java/com/huike/product/entity/Product.java
@@ -11,6 +11,7 @@ import java.math.BigDecimal;
 @Entity
 @Table(name = "tb_product")
 public class Product {
+
     @Id
     private Long id;
     private String productName;
