+++ product_service/src/main/java/com/huike/product/service/ProductService.java
@@ -1,17 +1,21 @@
 package com.huike.product.service;
 
 import com.huike.product.entity.Product;
-import com.huike.product.entity.ProductMy;
-import org.apache.ibatis.annotations.Param;
 
 public interface ProductService {
+
     Product findById(Long id);
 
-    ProductMy findByAdminIdAndPassword(Long id);
-
+    /**
+     * 保存
+     */
     void save(Product product);
-
+    /**
+     * 更新
+     */
     void update(Product product);
-
+    /**
+     * 删除
+     */
     void delete(Long id);
 }
