+++ product_service/src/main/java/com/huike/product/service/impl/ProductServiceImpl.java
@@ -1,10 +1,8 @@
 package com.huike.product.service.impl;
 
-import com.huike.product.dao.ProductDao;import com.huike.product.dao.ProductMapper;
+import com.huike.product.dao.ProductDao;
 import com.huike.product.entity.Product;
-import com.huike.product.entity.ProductMy;
 import com.huike.product.service.ProductService;
-import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,19 +11,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Autowired
     private ProductDao productDao;
-
-    @Autowired
-    private ProductMapper productMapper;
-
-    @Override
-    public ProductMy findByAdminIdAndPassword(Long id) {
-        return productMapper.findByAdminIdAndPassword(id);
-    }
     @Override
     public Product findById(Long id) {
         return productDao.findById(id).get();
     }
-
     /**
      * 保存
      *
@@ -33,7 +22,7 @@ public class ProductServiceImpl implements ProductService {
      */
     @Override
     public void save(Product product) {
-        productDao.save(product);
+         productDao.save(product);
     }
     /**
      * 更新
