--- product_service/src/main/resources/META-INF/maven/archetype.xml
@@ -1,9 +1,0 @@
-<archetype>
-  <id>product_service</id>
-  <sources>
-    <source>src/main/java/App.java</source>
-  </sources>
-  <testSources>
-    <source>src/test/java/AppTest.java</source>
-  </testSources>
-</archetype>
