--- product_service/src/main/resources/archetype-resources/src/main/java/App.java
@@ -1,13 +1,0 @@
-package $huike;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
