--- 覆盖数据库.bat
@@ -1,1 +1,0 @@
-mysqldump -uroot -p1234 shop > .\shop.sql
